Add files via upload
Project Structure:
1.main.cpp: Entry point of the application.
2.mainwindow.h: Header file for the main window class.
3.mainwindow.cpp: Implementation of the main window class.
4.mainwindow.ui: UI design file created using Qt Designer.

Running the Application
Open the Project in Qt Creator: Load the project files into Qt Creator.
Build the Project: Compile the project using the build tools provided by Qt Creator.
Run the Application: Execute the application to see the GUI for the half adder.

When the application runs, it will display a window with two combo boxes for binary input, and two labels to show the sum and carry. The outputs will update automatically when the inputs change.